fix bug in job script file

+++ batch/runbatch_extractLFP.sh
@@ -35,7 +35,10 @@ cd /u/home/x/xinniu/nwbPipeline/batch
 
 # make a copy of batch script so that we can submit jobs for other patients
 # when previous jobs are still running:
-cp runbatch_extractLFP.m runbatch_extractLFP_$JOB_ID.m
+if [ ! -f "runbatch_extractLFP_$JOB_ID.m" ]; then
+    echo "create job script: runbatch_extractLFP_$JOB_ID.m"
+    cp runbatch_extractLFP.m runbatch_extractLFP_$JOB_ID.m
+fi
 
 matlab  -nosplash -nodisplay -singleCompThread <<EOF
     runbatch_extractLFP_$JOB_ID($SGE_TASK_ID, $total_tasks);
